// src/app/core/models/userRoleDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface UserRoleDto {
  id: string;
  userId: string;
  roleId: string;
  roleName: string;
  createdAt: string;
  updatedAt: string;
}